requirements.txt
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
chromadb>=0.5.0
solidity-parser>=0.1.1
py-solc-x>=2.0.0
pydantic>=2.7.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
